Add auto select for variable attributes
Auto select now works on product page load, add to cart will save the selection of the user so auto select doesn't come into play when that button is hit. Add to cart is greyed out if no options are available. values that are not available are also greyed out.